refactor: update types for tags
