+++ src/core/features/tags/controller/TagsController.test.ts
@@ -1,7 +1,7 @@
 import { tmpdir } from 'os';
 import { tmpNameSync } from 'tmp';
 
-import { openDatabase } from '../../storage/database/SQLiteDatabase/SQLiteDatabase';
+import { openDatabase } from '../../../storage/database/SQLiteDatabase/SQLiteDatabase';
 
 import { TagsController } from './TagsController';
